Moved description check to Collector
