+++ spek-dsl/src/commonMain/kotlin/org/spekframework/spek2/lifecycle/Scope.kt
@@ -2,5 +2,4 @@ package org.spekframework.spek2.lifecycle
 
 interface Scope {
     val parent: GroupScope?
-    val id: ScopeId
 }
