+++ spek-runner-junit5/src/main/kotlin/org/spekframework/spek2/junit/SpekTestDescriptor.kt
@@ -5,15 +5,15 @@ import org.junit.platform.engine.TestSource
 import org.junit.platform.engine.TestTag
 import org.junit.platform.engine.UniqueId
 import org.junit.platform.engine.support.descriptor.ClassSource
-import org.spekframework.spek2.lifecycle.ScopeType
 import org.spekframework.spek2.runtime.scope.GroupScopeImpl
 import org.spekframework.spek2.runtime.scope.ScopeImpl
+import org.spekframework.spek2.runtime.scope.ScopeType
 import org.spekframework.spek2.runtime.scope.TestScopeImpl
 import java.util.*
 
 class SpekTestDescriptor internal constructor(
-    val scope: ScopeImpl,
-    private val factory: SpekTestDescriptorFactory
+        val scope: ScopeImpl,
+        private val factory: SpekTestDescriptorFactory
 ) : TestDescriptor {
 
     companion object {
