--- spek-runtime/src/commonMain/kotlin/org/spekframework/spek2/runtime/util/ValidationUtil.kt
@@ -1,6 +1,0 @@
-package org.spekframework.spek2.runtime.util
-
-import org.spekframework.spek2.lifecycle.ScopeId
-
-fun ScopeId.isValid(): Boolean =
-        name.isNotBlank() && !""" *\[\d*]""".toRegex().matches(name)
